percantik payroll

+++ app/Livewire/Prindexwr.php
@@ -8,19 +8,65 @@ use App\Models\Jamkerjaid;
 use Livewire\WithPagination;
 use App\Models\Yfrekappresensi;
 use Illuminate\Support\Facades\DB;
+use Livewire\Attributes\On;
 
 class Prindexwr extends Component
 {
     use WithPagination;
+    public $periode;
     public $search = '';
 
+
+    public function mount()
+    {
+        $getTglTerakhir = Jamkerjaid::select('date')
+            ->orderBy('date', 'desc')
+            ->first();
+        $this->periode = $getTglTerakhir->date;
+    }
+
+    public function rubah()
+    {
+        $this->render();
+    }
     public function updatingSearch()
     {
         $this->resetPage();
     }
 
+    public function getPayrollConfirmation()
+    {
+        $checkIfJamKerjaExist = Jamkerjaid::where('date', $this->periode)->first();
+        if ($checkIfJamKerjaExist) {
+            // $this->dispatch('error', message: 'Masih ada data no scan');
+            $this->dispatch('confirm');
+            return back();
+        } else {
+            $this->getPayroll();
+        }
+    }
+
+    #[On('getPayroll')]
     public function getPayroll()
     {
+        $checkIfJamKerjaExist = Jamkerjaid::where('date', $this->periode)->first();
+        if ($checkIfJamKerjaExist) {
+            $this->dispatch('error', message: 'Masih ada data no scan');
+            return back();
+        }
+
+        $tglsementara = Yfrekappresensi::where('no_scan', 'No Scan')
+            ->whereYear('date', getTahun($this->periode))
+            ->whereMonth('date', getBulan($this->periode))
+            ->count();
+
+        if ($tglsementara) {
+            $this->dispatch('error', message: 'Masih ada data no scan');
+            return back();
+        }
+
+        $checkIfJamKerjaExist = Jamkerjaid::where('date', $this->periode)->first();
+
         Jamkerjaid::query()->truncate();
         $jumlah_jam_terlambat = null;
         $jumlah_menit_lembur = null;
@@ -96,14 +142,15 @@ class Prindexwr extends Component
 
             $data = Jamkerjaid::find($data->id);
             $data->name = $dt->name;
+            $data->date = buatTanggal($dt->date);
             $data->jumlah_jam_kerja = $jumlah_jam_kerja;
             $data->jumlah_menit_lembur = $jumlah_menit_lembur;
             $data->jumlah_jam_terlambat = $total_late;
-            $data->first_in_late = $total_late_1;
-            $data->first_out_late = $total_late_2;
-            $data->second_in_late = $total_late_3;
-            $data->second_out_late = $total_late_4;
-            $data->overtime_in_late = $total_late_5;
+            $data->first_in_late = $total_late_1 == 0 ? null : $total_late_1;
+            $data->first_out_late = $total_late_2 == 0 ? null : $total_late_2;
+            $data->second_in_late = $total_late_3 == 0 ? null : $total_late_3;
+            $data->second_out_late = $total_late_4 == 0 ? null : $total_late_4;
+            $data->overtime_in_late = $total_late_5 == 0 ? null : $total_late_5;
             $data->save();
         }
         $this->dispatch('success', message: 'Data Karyawan Sudah di Save');
@@ -111,9 +158,18 @@ class Prindexwr extends Component
 
     public function render()
     {
-        $filteredData = Jamkerjaid::where('name', 'LIKE', '%' . trim($this->search) . '%')
+        // $getPeriode = Yfrekappresensi::distinct('date')->get();
+        $periodePayroll = DB::table('yfrekappresensis')
+            ->select(DB::raw('YEAR(date) year, MONTH(date) month, MONTHNAME(date) month_name'))
+            ->distinct()
+            ->orderBy('year', 'desc')
+            ->orderBy('month', 'desc')
+            ->get();
+
+        $filteredData = Jamkerjaid::whereDate('date', $this->periode)
+            ->orWhere('name', 'LIKE', '%' . trim($this->search) . '%')
             ->orWhere('user_id', 'LIKE', '%' . trim($this->search) . '%')
             ->paginate(10);
-        return view('livewire.prindexwr', compact(['filteredData']));
+        return view('livewire.prindexwr', compact(['filteredData', 'periodePayroll']));
     }
 }
